requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
lru-dict==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
parsimonious==0.9.0
prisma==0.11.0
protobuf==4.25.1
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
pyunormalize==15.1.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.12.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.3
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
web3==6.11.3
websockets==12.0
yarl==1.9.2
